@@ -10,21 +10,34 @@
     "lint": "eslint ."
   },
   "dependencies": {
-    "@react-native-async-storage/async-storage": "^1.17.7",
+    "@expo/vector-icons": "^13.0.0",
+    "@react-native-async-storage/async-storage": "^1.23.1",
+    "@react-native-community/geolocation": "2.1.0",
+    "@react-native-community/masked-view": "^0.1.11",
+    "@react-navigation/bottom-tabs": "^6.5.20",
+    "@react-navigation/drawer": "^6.6.15",
     "@react-navigation/native": "^6.0.11",
     "@react-navigation/stack": "^6.2.2",
+    "@reduxjs/toolkit": "^1.9.4",
     "axios": "^0.27.2",
-    "react": "17.0.2",
-    "react-native": "0.65.2",
-    "react-native-elements": "^3.4.2",
-    "react-native-gesture-handler": "^2.5.0",
+    "date-fns": "^2.29.0",
+    "expo-font": "^11.10.3",
+    "expo-modules-core": "^1.11.12",
+    "jsonwebtoken": "^9.0.2",
+    "react": "18.2.0",
+    "react-native": "0.68.0",
+    "react-native-elements": "^3.4.3",
+    "react-native-gesture-handler": "^2.16.0",
     "react-native-modal": "^13.0.1",
     "react-native-pager-view": "^5.4.25",
     "react-native-paper": "^4.12.3",
-    "react-native-reanimated": "^2.9.1",
-    "react-native-safe-area-context": "^4.3.1",
-    "react-native-screens": "^3.15.0",
+    "react-native-raw-bottom-sheet": "^3.0.0",
+    "react-native-reanimated": "^2.17.0",
+    "react-native-safe-area-context": "^4.9.0",
+    "react-native-screens": "^3.30.0",
+    "react-native-slider": "^0.11.0",
     "react-native-status-bar-height": "^2.6.0",
+    "react-native-swiper": "^1.6.0",
     "react-native-tab-view": "^3.1.1",
     "react-native-vector-icons": "^9.2.0",
     "react-navigation": "^4.4.4"
