@@ -1,4 +1,4 @@
 module.exports = {
   presets: ['module:metro-react-native-babel-preset'],
-  plugins: ['react-native-reanimated/plugin'],
+  plugins: [['react-native-reanimated/plugin', {relativeSourceLocation: true}]],
 };
